feat: move dot env to helper
