+++ packages/taro-cli/package.json
@@ -48,8 +48,6 @@
     "@tarojs/plugin-doctor": "0.0.10",
     "adm-zip": "^0.4.13",
     "cli-highlight": "^2.1.11",
-    "dotenv": "^16.0.3",
-    "dotenv-expand": "^9.0.0",
     "download-git-repo": "^2.0.0",
     "envinfo": "^7.8.1",
     "eslint": "^8.12.0",
