+++ packages/taro-helper/src/index.ts
@@ -5,6 +5,7 @@ export const createDebug = (id: string) => require('debug')(id)
 
 export { default as createBabelRegister, injectDefineConfigHeader } from './babelRegister'
 export * from './constants'
+export * from './dotenv'
 export * from './esbuild'
 export * as npm from './npm'
 export { default as createSwcRegister } from './swcRegister'
